requirements.txt
streamlit
zeep
requests
python-dotenv
google-generativeai
fpdf2
pandas